### 2023-04-26 #590 @salvafern
mregions2: Access Data from Marineregions.org: The Marine Regions Gazetteer and the Marine Regions Data Products
Submitting Author Name: Salvador Fernández Bejarano
Submitting Author Github Handle: @salvafern
Repository:  https://github.com/lifewatch/mregions2
Version submitted:
Submission type: Standard
Version accepted: TBD

---

```
Package: mregions2
Title: Access Data from Marineregions.org: The Marine Regions Gazetteer and the Marine Regions Data Products
Version: 1.0.0
  person(given = "Salvador", family = "Fernandez-Bejarano", role = c("aut", "cre"),
         email = "salvador.fernandez@vliz.be",
         comment = c(ORCID = "0000-0003-0535-7677", github = "salvafern")),
  person(given = "Lotte",family = "Pohl", role = "aut",
         email = "lotte.pohl@imbrsea.eu",
         comment = c(ORCID = "0000-0002-7607-7018", github = "whaleshark99")),
  person("LifeWatch Belgium", role = "fnd", comment = "https://lifewatch.be")
  )
Maintainer: Salvador Fernandez-Bejarano <salvador.fernandez@vliz.be>
Description: Explore and retrieve marine geo-spatial data from the Marine Regions Gazetteer and the Marine Regions Data Products, including the Maritime Boundaries.
    checkmate,
    glue,
    httr2,
    sf,
    rdflib,
    ISOcodes,
    memoise,
    cli,
    xml2,
    wrapr,
    curl,
    digest
    https://lifewatch.github.io/mregions2/
    ows4R (>= 0.3),
    httptest2,
    leaflet,
    leaflet.extras2,
    mapview,
    mregions,
    wk,
    withr

```

This package reads the [Marine Regions Gazetteer](https://marineregions.org/gazetteer.php) and the [Marine Regions Data Products](https://marineregions.org/sources.php) via a REST API and OGC web services respectively. The data is mostly geospatial as it contains locations of marine places in the world.

The main purpose of Marine Regions is to serve as the geographical backbone of biodiversity databases such as WoRMS or EurOBIS, hence the main user will be scientists that want to geo-reference their research. But it can also be useful for industry, policy and law makers.

There is already a package to read Marine Regions. The need for this second package is explained in one of the package's articles: [Why mregions2?](https://lifewatch.github.io/mregions2/articles/why_mregions2.html) . It was also discussed beforehand that a new package was the best approach: https://github.com/ropensci/mregions/pull/62 (Mostly to not break compatibility)

In general I think this does not apply. The only data being collected from users is the user-agent in all HTTP requests. This gets the environment variable `HTTPUserAgent` via `getOption()` in [mr_user_agent](https://github.com/lifewatch/mregions2/blob/9ed0de6db99b2415400ecc6f8d0df2f5b1577fef/R/utils.R#L48)

On the other hand, Marine Regions has sensitive data products such as the maritime boundaries. It is stated clearly on our [disclaimer](https://www.marineregions.org/disclaimer.php) that the data has no legal value whatsoever.
