### 2023-04-26 #590 @maurolepore
Thanks @salvafern for this submission and for pointing to https://github.com/ropensci/mregions/pull/62.

How can we leave a record that the maintainer of mregions is OK with the decision of submitting mregions2? Maybe you can point me to an GH-issue comment?
