### 2023-04-27 #590 @salvafern
Hi @maurolepore, thanks for checking out! 🚀

I am actually the current maintainer of mregions (https://github.com/ropensci/mregions/commit/e5fb45be93a957e4ccc51158bb900f72dbcd08a1). Before me were @LennertSchepers and @sckott (https://github.com/ropensci/mregions/commit/dd7b6e44aef74808f3f0f75a09d0568356ce632d).

Since we had this public conversation on https://github.com/ropensci/mregions/pull/62, and also got the idea of a new package from @maelle (see https://github.com/ropensci/mregions/pull/62#issuecomment-1061969979), I think we are all on the same page about submitting mregions2. But if there is any objection from Lennert or Scott I'm happy to discuss of course 😃
