### 2023-04-28 #590 @LennertSchepers
Worth noting if you need written approval: we are also the Marine Regions data system developers/maintainers/owners (see e.g. https://marineregions.org/editors.php)
