### 2023-05-05 #590 @jooolia
- Yes httptest2

    - .github/CONTRIBUTING.md exists, but only gives detail to email one of the authors and to abide by the Code of Conduct rules. I think this could be improved with resources contained here: https://devguide.ropensci.org/collaboration.html?q=contribu#contributing-guide

Github action checks are kept open as issues, one for each branch.
---

Thanks for your submission. Overall the package is in great shape. The only thing that I would note would that it would be even better with more information on how to contribute to the package. Feel free to let me know if there is a reason that additional info should not be added. I will start looking for reviewers.

Cheers, Julia

---
