### 2023-05-30 #590 @Kattuvan
@jooolia I accept your invitation to review it. This is my first attempt to review a package.
Do you have a Test Case   to look for bugs.
Do I have to get a data frame to test all those 30 functions/methods
