### 2023-05-30 #590 @Kattuvan
@jooolia I found the guidelines  https://devguide.ropensci.org/reviewerguide.html
I am sorry for my earlier mail
