### 2023-06-05 #590 @Kattuvan
Hi Julia,
Please find below the results I obtained using the 'pkgcheck'  on 'mregions2'.
Are you expecting me to use the package against the data for further check.
This is my first review of any package. I appreciate your guidance. Thanks
Murali.

> summary(x)

── mregions2 1.0.0 ──────────────────────────────────────────────────────────────

✔ Package name is available
✔ has a 'codemeta.json' file.
✔ has a 'contributing' file.
✔ uses 'roxygen2'.
✔ 'DESCRIPTION' has a URL field.
✔ 'DESCRIPTION' has a BugReports field.
✔ Package has at least one HTML vignette
✔ All functions have examples.
✖ Package has no continuous integration checks.
✔ Package coverage is 91.5%.
✖ R CMD check found 1 error.
✖ R CMD check found 1 warning.

ℹ Current status:
✖ This package is not ready to be submitted.

> print(x)

── mregions2 1.0.0 ──────────────────────────────────────────────────────────────

✔ Package name is available
✔ has a 'codemeta.json' file.
✔ has a 'contributing' file.
✔ uses 'roxygen2'.
✔ 'DESCRIPTION' has a URL field.
✔ 'DESCRIPTION' has a BugReports field.
✔ Package has at least one HTML vignette
✔ All functions have examples.
✖ Package has no continuous integration checks.
✔ Package coverage is 91.5%.
✖ R CMD check found 1 error.
✖ R CMD check found 1 warning.

ℹ Current status:
✖ This package is not ready to be submitted.

── git ──

• HEAD: 9ed0de6d
• Default branch: main
• Number of commits: 134
• First commit: 27-04-2022
• Number of authors: 3

── Package Structure ──

ℹ Package uses the following languages:
• : %

ℹ Package has
• 2 authors.
• 3 vignettes.
• 1 internal data file.
• 16 imported packages.
• 59 exported functions (median 3 lines of code).
• 123 non-exported functions (median 6 lines of code).
• 2 parameters per function (median).

── Package statistics ──

                   measure  value percentile noteworthy
1                  files_R    0.0        0.0       TRUE
4          files_vignettes    0.0        0.0       TRUE
5              files_tests    0.0        0.0       TRUE
6            num_vignettes    3.0       94.2
7          data_size_total 7328.0       69.2
8         data_size_median 7328.0       77.0
9                  n_fns_r  182.0       88.4
10        n_fns_r_exported   59.0       90.2
11    n_fns_r_not_exported  123.0       87.4
12        n_fns_per_file_r   10.2       87.7
13       num_params_per_fn    2.0       11.9
14            loc_per_fn_r    5.0        8.1
15        loc_per_fn_r_exp    3.0        1.5       TRUE
16    loc_per_fn_r_not_exp    6.0       13.8
17     doclines_per_fn_exp   55.0       68.4
18 doclines_per_fn_not_exp    0.0        0.0       TRUE
19    fn_call_network_size  100.0       79.2

ℹ Package network diagram is at [C:/Users/Murali/AppData/Local/Cache/R/pkgcheck/static/mregions2_pkgstats9ed0de6d.html].

── goodpractice ──

── GP mregions2 ─────────────────────────────────────────────────────────────────

It is good practice to

  ✖ write unit tests for all functions, and all package code in
    general. 91% of code lines are covered by test cases.

    R/00_gaz_geom.R:54:NA
    R/00_gaz_geom.R:338:NA
    R/00_gaz_geom.R:339:NA
    R/00_gaz_geom.R:340:NA
    R/00_gaz_geom.R:341:NA
    ... and 71 more lines

    data-raw\mrp_list.R:20:81
    R\00_gaz_geom.R:6:81
    R\00_gaz_geom.R:9:81
    R\00_gaz_geom.R:10:81
    R\00_gaz_geom.R:13:81
    ... and 387 more lines

  ✖ fix this R CMD check NOTE: Note: found 1 marked UTF-8
    string
  ✖ fix this R CMD check WARNING: LaTeX errors when creating
    PDF version. This typically indicates Rd problems.
─────────────────────────────────────────────────────────────────────────────────
── Package Versions ──

  pkgstats: 0.1.3.4
  pkgcheck: 0.1.1.24
> results_of_x <- print(x)

── mregions2 1.0.0 ──────────────────────────────────────────────────────────────

✔ Package name is available
✔ has a 'codemeta.json' file.
✔ has a 'contributing' file.
✔ uses 'roxygen2'.
✔ 'DESCRIPTION' has a URL field.
✔ 'DESCRIPTION' has a BugReports field.
✔ Package has at least one HTML vignette
✔ All functions have examples.
✖ Package has no continuous integration checks.
✔ Package coverage is 91.5%.
✖ R CMD check found 1 error.
✖ R CMD check found 1 warning.

ℹ Current status:
✖ This package is not ready to be submitted.

── git ──

• HEAD: 9ed0de6d
• Default branch: main
• Number of commits: 134
• First commit: 27-04-2022
• Number of authors: 3

── Package Structure ──

ℹ Package uses the following languages:
• : %

ℹ Package has
• 2 authors.
• 3 vignettes.
• 1 internal data file.
• 16 imported packages.
• 59 exported functions (median 3 lines of code).
• 123 non-exported functions (median 6 lines of code).
• 2 parameters per function (median).

── Package statistics ──

                   measure  value percentile noteworthy
1                  files_R    0.0        0.0       TRUE
4          files_vignettes    0.0        0.0       TRUE
5              files_tests    0.0        0.0       TRUE
6            num_vignettes    3.0       94.2
7          data_size_total 7328.0       69.2
8         data_size_median 7328.0       77.0
9                  n_fns_r  182.0       88.4
10        n_fns_r_exported   59.0       90.2
11    n_fns_r_not_exported  123.0       87.4
12        n_fns_per_file_r   10.2       87.7
13       num_params_per_fn    2.0       11.9
14            loc_per_fn_r    5.0        8.1
15        loc_per_fn_r_exp    3.0        1.5       TRUE
16    loc_per_fn_r_not_exp    6.0       13.8
17     doclines_per_fn_exp   55.0       68.4
18 doclines_per_fn_not_exp    0.0        0.0       TRUE
19    fn_call_network_size  100.0       79.2

ℹ Package network diagram is at [C:/Users/Murali/AppData/Local/Cache/R/pkgcheck/static/mregions2_pkgstats9ed0de6d.html].

── goodpractice ──

── GP mregions2 ─────────────────────────────────────────────────────────────────

It is good practice to

  ✖ write unit tests for all functions, and all package code in
    general. 91% of code lines are covered by test cases.

    R/00_gaz_geom.R:54:NA
    R/00_gaz_geom.R:338:NA
    R/00_gaz_geom.R:339:NA
    R/00_gaz_geom.R:340:NA
    R/00_gaz_geom.R:341:NA
    ... and 71 more lines

    data-raw\mrp_list.R:20:81
    R\00_gaz_geom.R:6:81
    R\00_gaz_geom.R:9:81
    R\00_gaz_geom.R:10:81
    R\00_gaz_geom.R:13:81
    ... and 387 more lines

  ✖ fix this R CMD check NOTE: Note: found 1 marked UTF-8
    string
  ✖ fix this R CMD check WARNING: LaTeX errors when creating
    PDF version. This typically indicates Rd problems.
─────────────────────────────────────────────────────────────────────────────────
── Package Versions ──

  pkgstats: 0.1.3.4
  pkgcheck: 0.1.1.24
    On Tuesday, May 30, 2023 at 11:37:37 p.m. EDT, Julia Gustavsen ***@***.***> wrote:

Thanks @sheilasaia for agreeing to review. Please feel free to reach out if you have any questions. Thanks, Julia
