### 2023-06-05 #590 @jooolia
Dear @Kattuvan , Great that you ran {pkgcheck}, will you follow up with the [review template](https://devguide.ropensci.org/reviewtemplate.html#reviewtemplate) ? Here there are specific things to check and an opportunity to provide your comments on the package to the authors (related to the code, the usability and functionality). Thanks, Julia
