### 2023-06-07 #590 @Kattuvan
Hi Julia,
Please find attached duly filled review template herewith.
Thank you.
Murali
    On Monday, June 5, 2023 at 02:16:43 a.m. EDT, Julia Gustavsen ***@***.***> wrote:

Dear @Kattuvan , Great that you ran {pkgcheck}, will you follow up with the review template ? Here there are specific things to check and an opportunity to provide your comments on the package to the authors (related to the code, the usability and functionality). Thanks, Julia
