### 2023-06-11 #590 @Kattuvan
## Package Review

-   ☒ As the reviewer I confirm that there are no [conflicts of interest](https://devguide.ropensci.org/policies.html#coi) for me to review this work (if you are unsure whether you are in conflict, please speak to your editor *before* starting your review).

#### Documentation

-   ☒ **A statement of need:** clearly stating problems the software is designed to solve and its target audience in README

-   ☒ **Installation instructions:** for the development version of package and any non-standard dependencies in README

-   dependencies ‘sf’; ‘dplyr’; ‘magritter’ are referenced elsewhere. <https://lifewatch.github.io/mregions2/articles/mregions2.html/>

-   ☒ **Vignette(s):** demonstrating major functionality that runs successfully locally

-   Package has at least one HTML vignette

-   But Vignette is available only through 'READ.md'

-   ☐ **Function Documentation:** for all exported functions

-   ☒ **Examples:** (that run successfully locally) for all exported functions

-   Examples don’t run in place.

-   All functions have examples

-   ☒ **Community guidelines:** including contribution guidelines in the README or CONTRIBUTING, and DESCRIPTION with `URL`, `BugReports` and `Maintainer` (which may be autogenerated via `Authors@R`).

-   availble

#### Functionality

-   ☒ **Installation:** Installation succeeds as documented.

-   Removed ‘sf’; ‘dplyr’; ‘magritter’ from local system to test installation of ‘mregion2’

-   I get error message “Error in loadNamespace(j &lt;- i\[\[1L\]\], c(lib.loc, .libPaths()), versionCheck = vI\[\[j\]\]) : there is no package called ‘magrittr’”

-   So, I installed ‘magrittr’ before attempting again ‘mregion2’

-   ‘mregions2’ installs dependencies ‘sf’ and ‘dplyr’

-   ☐ **Functionality:** Any functional claims of the software been confirmed.

-   ☐ **Performance:** Any performance claims of the software been confirmed.

-   ☒ **Automated tests:** Unit tests cover essential functions of the package and a reasonable range of inputs and conditions. All tests pass on the local machine.

-   It is good practice to write unit tests for all functions, and all package code in general. 91% of code lines are covered by test cases.

-   avoid long code lines, it is bad for readability. Also, many people prefer editor windows that are about 80 characters wide. Try make your lines shorter than 80 characters

-   ☒ **Packaging guidelines**: The package conforms to the rOpenSci packaging guidelines.

-   This package is not ready to be submitted.

-   Package has no continuous integration checks.

-   ☒ Should the author(s) deem it appropriate, I agree to be acknowledged as a package reviewer (“rev” role) in the package DESCRIPTION file.
