### 2023-06-11 #590 @Kattuvan
Hi Julia,
Done it as your requirements.
Murali
    On Friday, June 9, 2023 at 02:38:32 a.m. EDT, Julia Gustavsen ***@***.***> wrote:

Hi @Kattuvan, thanks for the doing the review. I think you may have attached something via email, but this did not appear on Github. Could you paste your review directly in the issue by logging into Github?
Thanks, Julia
