### 2023-06-25 #590 @sheilasaia
Here's my review.

---

Overall comments:

This package enables folks to directly access two specific datasets directly in their R session. I'm not familiar with the specific datasets myself, but can see how this would be helpful and a unique contribution to the R package space. I didn't have any problems with installing the package and all the lines of code in the vignette worked well including the leaflet maping functions. That said, there are several things that I think still need to be addressed before the package can be submitted.

I left some specific comments below. Thanks for the opportunity to review this rOpenSci package submission!

Some specific comments:

- I suggest you list the current maintainer on the GitHub page "About" section on the right (https://github.com/lifewatch/mregions2). I don't see that there, currently, but do see it in the DESCRIPTION file.
- I see there's an MIT licence file (license.md) given but this is not connected to the "About" section on the right of the main GitHub page. I suggest fixing the link to that file and also spelling out the specific author names directly in the license file so that's clear. It's currently listed as "authors".
- The code of conduct and citations files look complete.
- Can the authors give a little bit more info about the audience of this package in the description at the very top of the readme? It would be helpful to have a sentence or two at the very top that directly explains what types of data are in the Marine Regions Gazetteer and Marine Regions Data Products for those that are not familiar with these resources and would like to know this from quickly scanning the readme.
- Can the authors use the example in the R script for `gaz_search()`? Specifically it was `gaz_search("Belgian Part of the North Sea")`. When I first looked at this function in the vignette I mistook the input as having to only be a country so an example of what you all mean by "free text" would be helpful to see.
- When I run `gaz_rest()` I don't get a help page. I get an error that it cannot find the function: "Error in gaz_rest() : could not find function "gaz_rest"" I tried running without the `()` at the end and am also getting an error.
- In the usage description of `mrp_view()` please explain that all the functions under `mrp_view()` are for specific layers in the mrp data file. Maybe include a sub heading such as "Specific Layer Usage" or something similar to that to make this clear.
- The authors don't have a vignette for `mrp_col_unique()` but it would be helpful to include this since it can give the user a sense of what columns are available to them if they are not as familiar with the structure of the Marine Regions data product.
- The line `gaz_search_by_type("EEZ")` was taking me 13 seconds to run, which is still fast, but longer than the other functions were taking (i.e., generally most took 1-2 sections). I wanted to let the authors know this in case that's unexpected.
- It would be helpful to have a section in the bottom of the README that discusses whether there are similar packages or maybe even complimentary packges to this one. That way the user get get a quick background.
- According to the ropensci-review-bot, it looks like only 91% of the functions are covered by tests so I suggest the authors work on this so the package is 100% covered.
