### 2023-06-26 #590 @salvafern
Hi all,

Thank you very much @jooolia for coordinating the reviews; and @Kattuvan and @sheilasaia for your thorough reviews!

I'm currently the solely developer of mregions2 and I'm also going on leave soon till 2023-07-24, thus won't be able to address the reviews until then. Hope this is not a problem, otherwise I will try to find someone who can pick this up during my absence.
