+++ tests/testthat/prod/ok/geo/ows-bca3eb.R
@@ -2,5 +2,5 @@ structure(list(method = "GET", url = "geo/ows?service=wfs&version=2.0.0&request=
     status_code = 200L, headers = structure(list(`Access-Control-Allow-Origin` = "*", 
         `X-Frame-Options` = "SAMEORIGIN", `Content-Disposition` = "inline; filename=geoserver-GetPropertyValue.text", 
         `Content-Encoding` = "gzip", `Content-Type` = "text/xml", 
-        `Transfer-Encoding` = "chunked", Date = "Fri, 21 Apr 2023 11:59:02 GMT"), class = "httr2_headers"), 
+        `Transfer-Encoding` = "chunked", Date = "Wed, 28 Jun 2023 12:59:30 GMT"), class = "httr2_headers"), 
     body = charToRaw("<?xml version=\"1.0\" encoding=\"UTF-8\"?><wfs:ValueCollection xmlns:xs=\"http://www.w3.org/2001/XMLSchema\" xmlns:MarineRegions=\"geo.vliz.be/MarineRegions\" xmlns:fes=\"http://www.opengis.net/fes/2.0\" xmlns:wfs=\"http://www.opengis.net/wfs/2.0\" xmlns:gml=\"http://www.opengis.net/gml/3.2\" xmlns:ows=\"http://www.opengis.net/ows/1.1\" xmlns:xlink=\"http://www.w3.org/1999/xlink\" xmlns:xsi=\"http://www.w3.org/2001/XMLSchema-instance\" xsi:schemaLocation=\"http://www.opengis.net/wfs/2.0 geo/schemas/wfs/2.0/wfs.xsd\"><wfs:member><MarineRegions:doc_date>2018-12-07Z</MarineRegions:doc_date></wfs:member><wfs:member><MarineRegions:doc_date>2012-12-24Z</MarineRegions:doc_date></wfs:member><wfs:member><MarineRegions:doc_date>2014-12-15Z</MarineRegions:doc_date></wfs:member><wfs:member><MarineRegions:doc_date>2018-10-23Z</MarineRegions:doc_date></wfs:member><wfs:member><MarineRegions:doc_date>2012-04-19Z</MarineRegions:doc_date></wfs:member><wfs:member><MarineRegions:doc_date>2012-06-14Z</MarineRegions:doc_date></wfs:member><wfs:member><MarineRegions:doc_date>2018-04-06Z</MarineRegions:doc_date></wfs:member><wfs:member><MarineRegions:doc_date>2017-10-26Z</MarineRegions:doc_date></wfs:member><wfs:member><MarineRegions:doc_date>2009-03-31Z</MarineRegions:doc_date></wfs:member><wfs:member><MarineRegions:doc_date>2021-03-31Z</MarineRegions:doc_date></wfs:member><wfs:member><MarineRegions:doc_date>2017-08-22Z</MarineRegions:doc_date></wfs:member><wfs:member><MarineRegions:doc_date>2013-06-24Z</MarineRegions:doc_date></wfs:member><wfs:member><MarineRegions:doc_date>2009-05-12Z</MarineRegions:doc_date></wfs:member><wfs:member><MarineRegions:doc_date>2015-08-03Z</MarineRegions:doc_date></wfs:member><wfs:member><MarineRegions:doc_date>2019-04-11Z</MarineRegions:doc_date></wfs:member><wfs:member><MarineRegions:doc_date>2008-11-12Z</MarineRegions:doc_date></wfs:member><wfs:member><MarineRegions:doc_date>2010-12-02Z</MarineRegions:doc_date></wfs:member><wfs:member><MarineRegions:doc_date>2017-10-12Z</MarineRegions:doc_date></wfs:member><wfs:member><MarineRegions:doc_date>2017-01-18Z</MarineRegions:doc_d
[…]
