+++ tests/testthat/prod/ok/geo/ows-db8988.R
@@ -2,7 +2,7 @@ structure(list(method = "GET", url = "geo/ows?service=wfs&version=2.0.0&request=
     status_code = 200L, headers = structure(list(`Access-Control-Allow-Origin` = "*", 
         `X-Frame-Options` = "SAMEORIGIN", `Content-Disposition` = "inline; filename=MarineRegions-ecs_boundaries.xsd", 
         `Content-Encoding` = "gzip", `Content-Type` = "application/gml+xml; version=3.2", 
-        `Transfer-Encoding` = "chunked", Date = "Fri, 21 Apr 2023 11:58:59 GMT"), class = "httr2_headers"), 
+        `Transfer-Encoding` = "chunked", Date = "Wed, 28 Jun 2023 12:59:24 GMT"), class = "httr2_headers"), 
     body = as.raw(c(0x3c, 0x3f, 0x78, 0x6d, 0x6c, 0x20, 0x76, 
     0x65, 0x72, 0x73, 0x69, 0x6f, 0x6e, 0x3d, 0x22, 0x31, 0x2e, 
     0x30, 0x22, 0x20, 0x65, 0x6e, 0x63, 0x6f, 0x64, 0x69, 0x6e, 
