+++ pkgdown.yml
@@ -5,7 +5,7 @@ articles:
   mregions2-rdf: mregions2-rdf.html
   why_mregions2: why_mregions2.html
   mregions2: mregions2.html
-last_built: 2023-04-26T14:00Z
+last_built: 2023-06-28T13:22Z
 urls:
   reference: https://lifewatch.github.io/mregions2/reference
   article: https://lifewatch.github.io/mregions2/articles
