### 2023-09-07 c002464
docs: add rOpenSci status badge
