### 2023-09-07 6a7b244
docs: Update gaz_search() example
