+++ README.Rmd
@@ -15,25 +15,22 @@ knitr::opts_chunk$set(
 # mregions2
 
 <!-- badges: start -->
-[![Funding](https://img.shields.io/static/v1?label=powered+by&message=lifewatch.be&labelColor=1a4e8a&color=f15922)](https://lifewatch.be)
-[![Status at rOpenSci Software Peer Review](https://badges.ropensci.org/590_status.svg)](https://github.com/ropensci/software-review/issues/590)
-[![R-CMD-check](https://github.com/lifewatch/mregions2/actions/workflows/R-CMD-check.yaml/badge.svg)](https://github.com/lifewatch/mregions2/actions/workflows/R-CMD-check.yaml)
-[![pkgcheck](https://github.com/lifewatch/mregions2/workflows/pkgcheck/badge.svg)](https://github.com/lifewatch/mregions2/actions?query=workflow%3Apkgcheck)
-[![Codecov test
-coverage](https://codecov.io/gh/lifewatch/mregions2/branch/main/graph/badge.svg)](https://app.codecov.io/gh/lifewatch/mregions2?branch=main)
+
+[![Funding](https://img.shields.io/static/v1?label=powered+by&message=lifewatch.be&labelColor=1a4e8a&color=f15922)](https://lifewatch.be) [![Status at rOpenSci Software Peer Review](https://badges.ropensci.org/590_status.svg)](https://github.com/ropensci/software-review/issues/590) [![R-CMD-check](https://github.com/lifewatch/mregions2/actions/workflows/R-CMD-check.yaml/badge.svg)](https://github.com/lifewatch/mregions2/actions/workflows/R-CMD-check.yaml) [![pkgcheck](https://github.com/lifewatch/mregions2/workflows/pkgcheck/badge.svg)](https://github.com/lifewatch/mregions2/actions?query=workflow%3Apkgcheck) [![Codecov test coverage](https://codecov.io/gh/lifewatch/mregions2/branch/main/graph/badge.svg)](https://app.codecov.io/gh/lifewatch/mregions2?branch=main)
+
 <!-- badges: end -->
 
 mregions2 allows to access the [Marine Regions Gazetteer](https://marineregions.org/gazetteer.php) and the [Marine Regions Data Products](https://marineregions.org/sources.php) in R.
 
 You can find detailed information in the articles online:
 
-* [Introduction to mregions2](https://lifewatch.github.io/mregions2/articles/mregions2.html)
-* [Why mregions and mregions2?](https://lifewatch.github.io/mregions2/articles/why_mregions2.html)
-* [mregions2 as RDF](https://lifewatch.github.io/mregions2/articles/mregions2-rdf.html)
+-   [Introduction to mregions2](https://lifewatch.github.io/mregions2/articles/mregions2.html)
+-   [Why mregions and mregions2?](https://lifewatch.github.io/mregions2/articles/why_mregions2.html)
+-   [mregions2 as RDF](https://lifewatch.github.io/mregions2/articles/mregions2-rdf.html)
 
 ![](https://user-images.githubusercontent.com/54405067/156809171-1487bb9f-35af-4418-8e2f-93c24a59aad0.gif)
 
-<sup>(Source: https://www.vliz.be/en/news?p=show&id=8160)</sup>
+<sup>(Source: <https://www.vliz.be/en/news?p=show&id=8160>)</sup>
 
 ## Installation
 
@@ -45,6 +42,7 @@ devtools::install_github("lifewatch/mregions2")
 ```
 
 Load the library with
+
 ```{r, results='hide', warning=FALSE, message=FALSE}
 library(mregions2)
 ```
@@ -53,33 +51,38 @@ library(mregions2)
 
[…]
