+++ DESCRIPTION
@@ -34,7 +34,7 @@ Imports:
     methods,
     curl,
     digest
-RoxygenNote: 7.2.0
+RoxygenNote: 7.2.3
 URL: https://github.com/lifewatch/mregions2,
     https://lifewatch.github.io/mregions2/
 BugReports: https://github.com/lifewatch/mregions2/issues
