+++ vignettes/mregions2.Rmd
@@ -20,42 +20,45 @@ options(rdf_print_format = "turtle")
 options(rmarkdown.html_vignette.check_title = FALSE)
 ```
 
-[Marine Regions](https://marineregions.org/) is part of the [LifeWatch project](https://www.lifewatch.be/). Its activities focus on two main outputs: 
+`mregions2` offers a streamlined interface to access data from [Marine Regions](https://marineregions.org/) in R for researchers, marine scientists, and geospatial analysts seeking marine geographical information. [Marine Regions](https://marineregions.org/) is part of the [LifeWatch project](https://www.lifewatch.be/). Its activities focus on two main outputs:
 
-* The [Marine Regions Gazetteer](https://marineregions.org/gazetteer.php)
-* The [Marine Regions Data Products](https://marineregions.org/sources.php)
+-   The [Marine Regions Gazetteer](https://marineregions.org/gazetteer.php)
+-   The [Marine Regions Data Products](https://marineregions.org/sources.php)
 
 mregions2 allows to retrieve marine geospatial information from the Marine Regions Gazetteer and the Marine Regions Data Products in R.
 
 The **Marine Regions Gazetteer** is a standard list of marine georeferenced place names.
 
-> Gazetteer: a dictionary of geographical names. 
+> Gazetteer: a dictionary of geographical names.
 
-<sup>https://www.thefreedictionary.com/gazetteer</sup>
+<sup><https://www.thefreedictionary.com/gazetteer></sup>
 
 The entries in the Marine Regions Gazetteer are identified with an unique and persistent identifier, named the `?MRGID` (Marine Regions Gazetteer Unique Identifier)
 
 > Syntax `http://marineregions.org/mrgid/<number>`
 
-<sup>E.g. https://marineregions.org/mrgid/3293</sup>
+<sup>E.g. <https://marineregions.org/mrgid/3293></sup>
 
 In addition to the Marine Regions Gazetteer, the Marine Regions Team creates and hosts geographical Data Products, being the most popular one the [Marine Regions Maritime Boundaries](https://marineregions.org/eez.php).
 
 The mregions2 package aims to difference these two domains as clearly as possible, starting with the naming of its functions:
 
-* Gazetteer functions: `gaz_*()`
-* Data Products functions: `mrp_*()`
+-   Gazetteer functions: `gaz_*()`
+-   Data Products functions: `mrp_*()`
 
-This naming is not trivial, as mregions2 uses different methods to access the Marine Regions Gazetteer than the Marine Regions Data Products: A [RESTful API](https://www.marineregions.org/gazetteer.php?p=webservices) for the Marine Regions Gazetteer, and [OGC Web Services](https://www.marineregions.org/webservices.php) to get the Marine Regions Data Products. These are hosted at the [Flanders Marine Institute (VLIZ) geoserver](https://geo.vliz.be): 
+This naming is not trivial, as mregions2 uses different methods to access the Marine Regions Gazetteer than the Marine Regions Data Products: A [RESTful API](https://www.marineregions.org/gazetteer.php?p=webservices) for the Marine Regions Gazetteer, and [OGC Web Services](https://www.marineregions.org/webservices.php) to get the Marine Regions Data Products. These are hosted at the [Flanders Marine Institute (VLIZ) geoserver](https://geo.vliz.be):
 
 > `http://geo.vliz.be/`
 
 The data products can be visualized via [Web Map Services (WMS)](https://en.wikipedia.org/wiki/Web_Map_Service) with `mrp_view()` or downloaded by [Web Feature Services (WFS)](https://en.wikipedia.org/wiki/Web_Feature_Service) with `mrp_get()`.
 
[…]
