+++ README.Rmd
@@ -148,6 +148,45 @@ mrp_get("eez", path = "path/to/data")
 
 Get to know more in the [Get Started vignette](https://lifewatch.github.io/mregions2/articles/mregions2.html).
 
+```{r, include=FALSE, echo=FALSE}
+# Copy and paste output - do not include to not add more deps
+get_md <- function(pkgs){
+  get_titles <- function(x) packageDescription(x, fields = "Title")
+  titles <- purrr::map_chr(pkgs, get_titles)
+  cat(glue::glue("* [{pkgs}](https://docs.ropensci.org/{pkgs}): _{titles}_. \n"), sep = "\n")
+}
+```
+
+## Related packages
+
+### Marine biodiversity data
+
+The main purpose of Marine Regions is to serve as the geographical backbone for the World Register of Marine Species (WoRMS), an authoritative classification and catalogue of marine names. Here is a list of R packages to access Marine taxonomical and biogeographic data that can be combined with data from `mregions2`:
+
+```{r, include=FALSE, echo=FALSE}
+get_md(c("worrms", "eurobis", "robis", "rgbif"))
+# Change links manually
+```
+* [worrms](https://docs.ropensci.org/worrms): _World Register of Marine Species (WoRMS) Client_. 
+* [eurobis](https://lifewatch.github.io/eurobis/): _Download data from EurOBIS using the LifeWatch/EMODnet-Biology Web Feature Services_. 
+* [robis](https://cran.r-project.org/package=robis): _Ocean Biodiversity Information System (OBIS) Client_. 
+* [rgbif](https://docs.ropensci.org/rgbif): _Interface to the Global Biodiversity Information Facility API_. 
+
+### More geographical data
+
+There are other R packages that allow to access other gazetteers and boundaries data. Here is a non-exhaustive list:
+
+```{r, include=FALSE, echo=FALSE}
+get_md(c("geonames", "osmdata", "osmextract", "rnaturalearth"))
+```
+
+* [geonames](https://docs.ropensci.org/geonames): _Interface to the "Geonames" Spatial Query Web Service_. 
+* [osmdata](https://docs.ropensci.org/osmdata): _Import 'OpenStreetMap' Data as Simple Features or Spatial
+Objects_. 
+* [osmextract](https://docs.ropensci.org/osmextract): _Download and Import Open Street Map Data Extracts_. 
+* [rnaturalearth](https://docs.ropensci.org/rnaturalearth): _World Map Data from Natural Earth_.
+
+
 ## Citation
 
 ```{r citation, cache=FALSE}
