### 2023-09-08 de65fbc
docs: Write better contributing guidelines
