+++ .github/CONTRIBUTING.md
@@ -1,7 +1,134 @@
+---
+editor_options: 
+  markdown: 
+    wrap: sentence
+---
+
 # Contributing to mregions2
 
-If you wish to contribute to mregions2 please get in contact with the team by email: <info@marineregions.org>
+```{=html}
+<!-- This CONTRIBUTING.md is adapted from:
+https://gist.github.com/peterdesmet/e90a1b0dc17af6c12daf6e8b2f044e7c 
+https://github.com/ropensci/skimr/blob/main/.github/CONTRIBUTING.md#understanding-the-scope-of-skimr
+usethis::use_tidy_contributing()
+-->
+```
+First of all, thanks for considering contributing to mregions2!
+mregions2 is an open source project maintained by the data managers of [Marine Regions](marineregions.org/), who are quite busy updating maritime boundaries and curating the database and really appreciate any help.
+
+## How you can contribute
+
+There are several ways you can contribute to this project.
+If you want to know more about why and how to contribute to open source projects like this one, see this [Open Source Guide](https://opensource.guide/how-to-contribute/).
+
+### Share the love ❤️
+
+Think mregions2 is useful?
+Let others discover it, by telling them in person, via Twitter or a blog post.
+
+Using mregions2 for a paper you are writing?
+Consider [citing it](https://lifewatch.github.io/mregions2/authors.html).
+
+### Ask a question ⁉️
+
+Using mregions2 and got stuck?
+Browse the [documentation](https://lifewatch.github.io/mregions2) to see if you can find a solution.
+Still stuck?
+Post your question as an [issue on GitHub](https://github.com/lifewatch/mregions2/issues/new).
+While we cannot offer user support, we'll try to do our best to address it, as questions often lead to better documentation or the discovery of bugs.
+
+Want to ask a question in private?
+Contact Marine Regions by [email](mailto:info@marineregions.org).
+
+### Propose an idea 💡
+
+Have an idea for a new mregions2 feature?
+Take a look at the [documentation](https://lifewatch.github.io/mregions2) and [issue list](https://github.com/lifewatch/mregions2/issues) to see if it isn't included or suggested yet.
+If not, suggest your idea as an [issue on GitHub](https://github.com/lifewatch/mregions2/issues/new).
+While we can't promise to implement your idea, it helps to: \* Explain in detail how it would work.
+\* Keep the scope as narrow as possible.
+
+See below if you want to contribute code for your idea as well.
+
+### Report a bug 🐛
+
+Using mregions2 and discovered a bug?
+That's annoying!
+Don't let others have the same experience and report it as an [issue on GitHub](https://github.com/lifewatch/mregions2/issues/new) so we can fix it.
+A good bug report makes it easier for us to do so.
+Please include:
+
+-   Output of `sessionInfo()`
+
+-   A minimal example made with [reprex](https://reprex.tidyverse.org/).
+
+-   Any details about your local setup that might be helpful in troubleshooting.
+
+### Improve the documentation 📖
+
+Noticed a typo on the website?
+Think a function could use a better example?
+Good documentation makes all the difference, so your help to improve it is very welcome!
+
+#### The website
+
+[This website](https://lifewatch.github.io/mregions2) is generated with [`pkgdown`](http://pkgdown.r-lib.org/).
+That means we don't have to write any html: content is pulled together from documentation in the code, vignettes, [Markdown](https://guides.github.com/features/mastering-markdown/) files, the package `DESCRIPTION` and `_pkgdown.yml` settings.
+If you know your way around `pkgdown`, you can [propose a file change](https://help.github.com/articles/editing-files-in-another-user-s-repository/) to improve documentation.
+If not, [report an issue](https://github.com/lifewatch/mregions2/issues/new) and we can point you in the right direction.
+
+#### Function documentation
+
[…]
