### 2023-09-12 dcbd99e
docs: make build_vignettes = T and load magrittr in README
