### 2023-09-13 #590 @salvafern
Hi @jooolia

Yes I am currently addressing the reviews (see https://github.com/lifewatch/mregions2/issues/1#issuecomment-1710318418 , will be updating the comment).
Apologies for the delay, I haven't been able to allocate time before. I should be able to do most of the changes this week.
