### 2023-09-15 lifewatch/mregions2#20 @salvafern
List definitions of each attribute of each data product
Raised by @sheilasaia at https://github.com/ropensci/software-review/issues/590#issuecomment-1606138219

> _The authors don't have a vignette for mrp_col_unique() but it would be helpful to include this since it can give the user a sense of what columns are available to them if they are not as familiar with the structure of the Marine Regions data product [...]_

This is challenging for us, the Marine Regions Team, as the attributes of the data products don't follow any convention. Plus, in some cases none of us created the data products nor decided how to name or what attributes should be there. Nonetheless, we will add an exhaustive list defining each attributes the best we can to our understanding. Future efforts shall include standardization of such attributes, potentially reusing on-going standards.
