+++ README.md
@@ -32,6 +32,8 @@ You can find detailed information in the articles online:
   mregions2](https://lifewatch.github.io/mregions2/articles/mregions2.html)
 - [Why mregions and
   mregions2?](https://lifewatch.github.io/mregions2/articles/why_mregions2.html)
+- [Marine Regions Data Products
+  Ontology](https://lifewatch.github.io/mregions2/articles/mrp_ontology.html)
 - [mregions2 as
   RDF](https://lifewatch.github.io/mregions2/articles/mregions2-rdf.html)
 
@@ -189,10 +191,9 @@ gaz_search(3293) %>% gaz_geometry()
 #>   MRGID gazetteerSource    placeType latitude longitude minLatitude minLongitude
 #>   <int> <chr>              <chr>        <dbl>     <dbl>       <dbl>        <dbl>
 #> 1  3293 Flanders Marine I… EEZ           51.5      2.71        51.1         2.24
-#> # … with 8 more variables: maxLatitude <dbl>, maxLongitude <dbl>,
-#> #   precision <dbl>, preferredGazetteerName <chr>,
-#> #   preferredGazetteerNameLang <chr>, status <chr>, accepted <int>,
-#> #   the_geom <MULTIPOLYGON [°]>
+#> # ℹ 8 more variables: maxLatitude <dbl>, maxLongitude <dbl>, precision <dbl>,
+#> #   preferredGazetteerName <chr>, preferredGazetteerNameLang <chr>,
+#> #   status <chr>, accepted <int>, the_geom <MULTIPOLYGON [°]>
 
 # Or get only the geometry 
 gaz_geometry(3293, format = "sfc")
@@ -211,7 +212,7 @@ hierarchically. You can browse this hierarchy up and down with
 ``` r
 # Get all relations
 gaz_search(3293) %>% gaz_relations()
-#> # A tibble: 31 × 14
+#> # A tibble: 36 × 14
 #>    MRGID gazetteerSource   placeType latitude longitude minLatitude minLongitude
 #>    <int> <chr>             <chr>        <dbl>     <dbl>       <dbl>        <dbl>
 #>  1  3293 Flanders Marine … EEZ           51.5      2.71        51.1         2.24
@@ -224,13 +225,14 @@ gaz_search(3293) %>% gaz_relations()
 #>  8  2550 ASFA thesaurus    Coast         51.2      2.91        51.1         2.54
 #>  9 17401 Aphia             Wreck         51.4      2.32        NA          NA   
 #> 10 17409 Aphia             Wreck         51.1      2.33        NA          NA   
-#> # … with 21 more rows, and 7 more variables: maxLatitude <dbl>,
-#> #   maxLongitude <dbl>, precision <dbl>, preferredGazetteerName <chr>,
-#> #   preferredGazetteerNameLang <chr>, status <chr>, accepted <int>
+#> # ℹ 26 more rows
+#> # ℹ 7 more variables: maxLatitude <dbl>, maxLongitude <dbl>, precision <dbl>,
+#> #   preferredGazetteerName <chr>, preferredGazetteerNameLang <chr>,
+#> #   status <chr>, accepted <int>
 
 # Or get the relations directly
 gaz_relations(3293)
-#> # A tibble: 31 × 14
+#> # A tibble: 36 × 14
 #>    MRGID gazetteerSource   placeType latitude longitude minLatitude minLongitude
 #>    <int> <chr>             <chr>        <dbl>     <dbl>       <dbl>        <dbl>
 #>  1  3293 Flanders Marine … EEZ           51.5      2.71        51.1         2.24
@@ -243,9 +245,10 @@ gaz_relations(3293)
 #>  8  2550 ASFA thesaurus    Coast         51.2      2.91        51.1         2.54
[…]
