### 2023-09-19 lifewatch/mregions2#21 @salvafern
CI fails due to deps of rdflib
Something is going on with the system dependencies of ropensci/rdflib on linux. It worked fine 3 months ago and now it is failing. e.g.

https://github.com/lifewatch/mregions2/actions/runs/6229986973/job/16909345713#step:5:5614
`libraptor2-dev : Depends: libcurl4-gnutls-dev but it is not installable`

https://github.com/lifewatch/mregions2/actions/runs/6198673859/job/16829577019
```
  The following packages have unmet dependencies:
   librdf0-dev : Depends: libraptor2-dev (>= 2.0.10) but it is not going to be installed
                 Depends: librasqal3-dev (>= 0.9.30) but it is not going to be installed
```

I have tried installing the deps apart but it is not solving it. Problem arises at the github action either https://github.com/r-lib/actions/tree/v2/setup-r-dependencies or https://github.com/ropensci-review-tools/pkgcheck-action

I can see the issue is also happening now on the ropensci/rdflib side, but it has not been tackled:
https://github.com/ropensci/rdflib/actions/runs/6193206211/job/16814374463#step:5:3545

There is a possibility that this gets solves alone after some time... Don't really know what is going on
