### 2023-09-20 lifewatch/mregions2#21 @salvafern
Won't fix as most likely it will be fix alone in a future version of `pak` -> https://github.com/ropensci/rdflib/issues/47#issuecomment-1725934662
