### 2023-09-20 b678e70
docs: add hex logo
