+++ README.Rmd
@@ -13,7 +13,8 @@ knitr::opts_chunk$set(
 )
 ```
 
-# mregions2
+# mregions2 <a href="https://lifewatch.github.io/mregions2/"><img src="man/figures/logo.png" align="right" height="137" alt="mregions2 website" /></a>
+
 
 <!-- badges: start -->
 
