### 2023-09-21 lifewatch/mregions2#20 @salvafern
The list with definitions is here: https://github.com/lifewatch/mregions2/blob/main/inst/mrp_ontology.csv

This is provisional and we may change it.
