### 2023-10-11 #590 @sheilasaia
@salvafern, thanks for your replies! I will check back at my `gaz_rest` comment and get back to you with session info, hopefully by the end of the weekend.
