### 2023-10-23 a86a1d4
docs: improve description gaz_rest (#1)
