### 2023-10-23 #590 @salvafern
Hi @sheilasaia, Indeed it was not that clear. The purpose of `gaz_rest` was to explain quickly why there are functions with names starting as `gaz_rest_*`. I think it is beneficial to keep this doc as I can easily link it to all `gaz_rest_*` functions through the section @ seealso  . I improved the explanations now (https://github.com/lifewatch/mregions2/commit/a86a1d4ad391bfb96d7f29171056fecda0568779). Let me know if it is better or you would still remove it from the docs.
