### 2024-03-10 #590 @jooolia
Dear @salvafern,
Things are looking good, although the comments about the CI have not fully been addressed in my view. I had a quick look, but the logs are too old (more than three months) to be viewed. Could you try re-running the Github Actions for the pkg-check and R-CMD-check so that we can help troubleshoot what is happening here?

Thanks!
