### 2024-04-09 #590 @salvafern
Hi @jooolia, apologies for the delay. The logs are fresh now: https://github.com/lifewatch/mregions2/actions/runs/8615096758/job/23609894330

I think the same issue continues: See https://github.com/lifewatch/mregions2/issues/21 and https://github.com/ropensci/rdflib/issues/47

Any ideas are highly appreciated!
