### 2024-04-28 #590 @jooolia
Thanks @salvafern for the reruns. I will have a look and get back to you (and it seems that rdflib has commented out the ubuntu tests for now). thanks, Julia
