+++ .github/workflows/pkgcheck.yaml
@@ -12,12 +12,12 @@ on:
   push:
     branches:
       - main
-      - dev
 
 jobs:
   pkgcheck:
+    env:
+        PKG_SYSREQS: false 
     runs-on: ubuntu-latest
     steps:
-      - name: Update apt-get cache
-        run: sudo apt-get update
+      - run: sudo apt-get install -y make libgdal-dev gdal-bin libgeos-dev libproj-dev libsqlite3-dev libssl-dev make zlib1g-dev libsecret-1-dev pandoc libpng-dev librdf0-dev libsodium-dev libicu-dev libfontconfig1-dev libfreetype6-dev libudunits2-dev libxml2-dev libjpeg-dev libtiff-dev libfribidi-dev libharfbuzz-dev libcurl4-gnutls-dev
       - uses: ropensci-review-tools/pkgcheck-action@main
