### 2024-05-07 #590 @jooolia
Hi @salvafern, I have gotten R-CMD-check to run without errors, although it is a bit of a workaround at the moment. I will look into the yaml files tonight and update you then (or if all goes well submit a pull request) with what has worked and what we could imagine doing.
Cheers, Julia
