### 2024-05-20 f8a8d4e
removed pkgcheck
Update DESCRIPTION removed extra sys requirements

 Removed testing info.
