+++ .github/workflows/pkgcheck.yaml
@@ -1,23 +1,0 @@
-name: pkgcheck
-
-# This will cancel running jobs once a new run is triggered
-concurrency:
-  group: ${{ github.workflow }}-${{ github.head_ref }}
-  cancel-in-progress: true
-
-on:
-  # Manually trigger the Action under Actions/pkgcheck
-  workflow_dispatch:
-  # Run on every push to main
-  push:
-    branches:
-      - main
-
-jobs:
-  pkgcheck:
-    env:
-        PKG_SYSREQS: false 
-    runs-on: ubuntu-latest
-    steps:
-      - run: sudo apt-get install -y make libgdal-dev gdal-bin libgeos-dev libproj-dev libsqlite3-dev libssl-dev make zlib1g-dev libsecret-1-dev pandoc libpng-dev librdf0-dev libsodium-dev libicu-dev libfontconfig1-dev libfreetype6-dev libudunits2-dev libxml2-dev libjpeg-dev libtiff-dev libfribidi-dev libharfbuzz-dev libcurl4-gnutls-dev
-      - uses: ropensci-review-tools/pkgcheck-action@main
