### 2024-05-20 lifewatch/mregions2#24 @jooolia
Github actions, some potential workarounds
Dear @salvafern ,
As discussed in https://github.com/ropensci/software-review/issues/590 there have been some difficulties with the testing of the package on Github Actions.

In my fork of the repo I have found a few workaround and I will make a pull request based on this work. We can discuss further in the PR.

Thanks, Julia
