### 2024-05-20 #590 @jooolia
Hi @salvafern I have opened a PR in the repo to your dev branch. Let me know what you think. I had to remove pkgcheck, but all three others I was able to get to work and pass.
Cheers, Julia
