### 2024-05-22 lifewatch/mregions2#24 @mpadge
@jooolia @salvafern The pkgcheck-action failures look like a cache issue to me, and I can't repeat in the container itself. I've nevertheless updated our Docker container in the commit linked above which should fix, so you can leave the `pkgcheck.yaml` file in place.
