### 2024-05-22 lifewatch/mregions2#24 @jooolia
Hi @mpadge, thanks for the check. I have tried to test this out but did not have any luck (you can see the run here: https://github.com/jooolia/mregions2/actions/runs/9196705137/job/25295406943), any thoughts on what I missed?
Thanks, Julia
