### 2024-05-23 lifewatch/mregions2#24 @mpadge
Ah, thanks @jooolia, that was enough to diagnose the problem at least. Can you please trigger the workflow again on your fork? Hopefully it should work now ... :crossed_fingers:
