### 2024-05-24 lifewatch/mregions2#24 @jooolia
Hi @mpadge, thanks, it is still failing. It seems to be similar to the failure with R-CMD-CHECK where we have the dependency temporary workaround in place.
