### 2024-06-03 5fc28d6
fix: geometries retrieval now takes into account attributeValue, fixes #26
