### 2024-06-03 lifewatch/mregions2#20 @salvafern
I will close for now as there are no immediate plans to make a better fix than what is currently on place
