### 2024-06-03 lifewatch/mregions2#24 @salvafern
Hi @jooolia and @mpadge , thanks a lot for having a look at this issue! As in https://github.com/ropensci-review-tools/pkgcheck-action/issues/48#issuecomment-2134551849 I assume this issue can be closed as the problem should be fixed at `pak`. For the time being  @jooolia 's PR seems to fix the CI. Feel free to reopen otherwise!
