### 2024-06-09 #590 @ropensci-review-bot
Approved! Thanks @salvafern for submitting and @Kattuvan, @sheilasaia for your reviews! :grin:
